Added in an error message for treeIO
All of my troubles recently have been caused by the input file
parser dying silently and thus appearing as if nothing was rendered.
This commit adds in a QMessageDialog after the read tree function
and checks to see whether the parser returned false or is there
are no leaves in the resulting tree. If either of these conditions
is true then an error message appears and no rendering is attemped