fix bug in rerooting tree
The x position of nodes was being set incorrectly as when the
new root node is created the height parameter was being set to -1
which had some streange flow on effects such that everthing was being
painted off the viewport.

No idea why this was happening with the ported code as the rerooting
functionality has nothing to do with the GUI. Anyway, the solution
was to recalculate the statistics for the whole tree again after
rerooting. This may not be efficient but hopfully rerooting will only
occur occassionally.